Unify StreamJunction impl and fix StateEvent handling
