+++ src/core/stream/mod.rs
@@ -4,7 +4,6 @@ pub mod handler;
 pub mod input;
 pub mod junction_factory;
 pub mod mapper;
-pub mod optimized_stream_junction;
 pub mod output;
 pub mod stream_initializer;
 pub mod stream_junction;
@@ -13,11 +12,7 @@ pub use self::handler::{SinkStreamHandler, SourceStreamHandler};
 pub use self::input::source::{timer_source::TimerSource, Source};
 pub use self::input::{InputHandler, InputManager};
 pub use self::junction_factory::{
-    BenchmarkResult, JunctionBenchmark, JunctionConfig, JunctionType, PerformanceLevel,
-    StreamJunctionFactory,
-};
-pub use self::optimized_stream_junction::{
-    JunctionPerformanceMetrics, OptimizedPublisher, OptimizedStreamJunction,
+    BenchmarkResult, JunctionBenchmark, JunctionConfig, StreamJunctionFactory,
 };
 pub use self::output::{LogSink, Sink, StreamCallback};
 pub use self::stream_initializer::{
@@ -25,7 +20,7 @@ pub use self::stream_initializer::{
     InitializedStreamHandler, StreamHandlers,
 };
 pub use self::stream_junction::{
-    OnErrorAction, Publisher, Receiver as StreamJunctionReceiver, StreamJunction,
+    JunctionPerformanceMetrics, OnErrorAction, Publisher, StreamJunction,
 };
 
 // Re-export mapper types for convenience
