+++ tests/error_routing.rs
@@ -103,18 +103,18 @@ fn test_fault_stream_routing() {
         "Fault".to_string(),
         Arc::clone(&fault_def),
         Arc::new(app_ctx.clone()),
-        8,
+        64,
         false,
         None,
-    )));
+    ).unwrap()));
     let mut main_junction = StreamJunction::new(
         "In".to_string(),
         Arc::clone(&def),
         Arc::new(app_ctx.clone()),
-        8,
+        64,
         true,
         Some(Arc::clone(&fault_junction)),
-    );
+    ).unwrap();
     main_junction.set_on_error_action(OnErrorAction::STREAM);
     let rec = Arc::new(Mutex::new(Vec::new()));
     fault_junction
@@ -149,10 +149,10 @@ fn test_error_store_routing() {
         "In".to_string(),
         Arc::clone(&def),
         Arc::new(app_ctx),
-        8,
+        64,
         true,
         None,
-    );
+    ).unwrap();
     junction.set_on_error_action(OnErrorAction::STORE);
     junction.stop_processing();
     junction.send_event(Event::new_with_data(0, vec![AttributeValue::Int(2)]));
