+++ tests/join_queries.rs
@@ -68,7 +68,7 @@ fn setup_context() -> (
         1024,
         false,
         None,
-    )));
+    ).unwrap()));
     let right_junction = Arc::new(Mutex::new(StreamJunction::new(
         "RightStream".to_string(),
         Arc::clone(&right_def),
@@ -76,7 +76,7 @@ fn setup_context() -> (
         1024,
         false,
         None,
-    )));
+    ).unwrap()));
     let out_junction = Arc::new(Mutex::new(StreamJunction::new(
         "OutStream".to_string(),
         Arc::clone(&out_def),
@@ -84,7 +84,7 @@ fn setup_context() -> (
         1024,
         false,
         None,
-    )));
+    ).unwrap()));
 
     let mut map = HashMap::new();
     map.insert("LeftStream".to_string(), left_junction);
