+++ tests/stream_junction_stress.rs
@@ -107,12 +107,15 @@ fn setup_junction(
         4096,
         async_mode,
         None,
-    )));
+    ).unwrap()));
     let rec = Arc::new(Mutex::new(Vec::new()));
     let p = Arc::new(Mutex::new(RecordingProcessor {
         events: Arc::clone(&rec),
     }));
     junction.lock().unwrap().subscribe(p);
+    if async_mode {
+        junction.lock().unwrap().start_processing().expect("Failed to start async processing");
+    }
     (junction, rec)
 }
